pyproject.toml
[project]
name = "langgraph-practice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langfuse>=4.0.6",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
